git-svn-id: http://svn.resourcespace.com/svn/rs/branches/20210723_abaxter_q11889@18250 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ plugins/csv_upload/include/csv_functions.php
@@ -454,7 +454,7 @@ function csv_upload_process($filename,&$meta,$resource_types,&$messages,$csv_set
                 {
                 if(!isset($newref))
                     {
-                    $lastref = sql_value("SELECT MAX(ref) value FROM resource",0);
+                    $lastref = (int) sql_value("SELECT MAX(ref) value FROM resource",0);
                     $newref  = $lastref + 1;
                     }
                 else
